requirements.txt
gradio_client==1.8.0
python-telegram-bot==21.4
python-dotenv==0.19.0
